requirements.txt
aiogram==3.5.0
python-dotenv==1.0.1
supabase==2.6.0